refactor: realign app views/serializers with models
models are kept in place for now and related code was moved to be in the
same app as the model it references

the only remaining issue is that core.views imports library.models.Codebases
in order to add related codebases to the context for MemberProfileViewSet

- removed some lingering references to the now-deprecated Institution
  model
- some additional splitting/cleaning up for readability

ref comses/planning#136
