+++ django/core/jinja2/account/login.jinja
@@ -21,7 +21,7 @@ Use the social sign in buttons below to sign in to CoMSES with your <a href='htt
 or <a href='https://orcid.org'>ORCID</a> account. If you don't have an existing CoMSES Net account this will create
 a new account and prompt you for additional information. If you <em>do</em> have an existing CoMSES Net account and want
 to link it with GitHub or ORCID, first sign in using your username / password below, then visit
-<a href='{{url("home:account-profile")}}#local-signin'>your account profile page</a> to link your accounts.
+<a href='{{url("core:account-profile")}}#local-signin'>your account profile page</a> to link your accounts.
 </p>
 {% include "includes/socialauth_login.jinja" %}
 <hr>
