+++ django/core/jinja2/core/jobs/list.jinja
@@ -23,7 +23,7 @@
                 {% endif %}
                 <small class="text-muted text-end">Last updated {{ timesince(item.last_modified) }} ago</small>
                 <small>
-                    Submitted by {{ member_profile_href(item.submitter) }}
+                    Submitted by {{ member_profile_href(item.submitter.member_profile) }}
                 </small>
             </div>
         </div>
@@ -49,7 +49,7 @@
 
 {% block introduction %}<h1>Jobs &amp; Appointments </h1>{% endblock %}
 
-{% set pagination_block = paginator('home:job-list', __all__) %}
+{% set pagination_block = paginator('core:job-list', __all__) %}
 
 {% block top %}
 {{ breadcrumb([
@@ -60,12 +60,12 @@
 {% endblock %}
 
 {% block search %}
-{{ search_bar(url("home:job-list"), "Search Jobs") }}
+{{ search_bar(url("core:job-list"), "Search Jobs") }}
 {{ pagination_block }}
 {% endblock %}
 
 {% block content %}
-{{ list_page(__all__, render_job, 'home:job-detail') }}
+{{ list_page(__all__, render_job, 'core:job-detail') }}
 {% if count == 0 %}
 <div class='mt-3 alert alert-primary'>
     <i class='fas fa-robot fa-2x'></i>
