+++ django/home/jinja2/home/index.jinja
@@ -87,7 +87,7 @@
                 </div>
             </div>
             <div class='col'>
-                <a href='{{ url("home:event-list") }}'>
+                <a href='{{ url("core:event-list") }}'>
                     <img src='{{ static("images/icons/groups.png") }}' alt='Upcoming Events'>
                     <h3>Upcoming Events</h3>
                 </a>
@@ -183,7 +183,7 @@
                         <h3>Latest Jobs &amp; Appointments</h3>
                     </div>
                     <div class='col-2 text-end'>
-                        <small><a href='{{ url("home:job-list") }}'>View All</a></small>
+                        <small><a href='{{ url("core:job-list") }}'>View All</a></small>
                     </div>
                 </div>
                 {% for item in latest_jobs %}
