+++ django/library/views.py
@@ -20,7 +20,6 @@ from ipware import get_client_ip
 from rest_framework import (
     viewsets,
     generics,
-    renderers,
     status,
     permissions,
     filters,
@@ -39,15 +38,15 @@ from rest_framework.response import Response
 from core.models import MemberProfile
 from core.permissions import ViewRestrictedObjectPermissions
 from core.view_helpers import add_change_delete_perms, get_search_queryset
+from core.mixins import CommonViewSetMixin
 from core.views import (
-    CommonViewSetMixin,
     FormUpdateView,
     FormCreateView,
-    SmallResultSetPagination,
     NoDeleteViewSet,
     NoDeleteNoUpdateViewSet,
     HtmlNoDeleteViewSet,
 )
+from core.pagination import SmallResultSetPagination
 from core.serializers import RelatedMemberProfileSerializer
 from .forms import (
     PeerReviewerFeedbackReviewerForm,
